requirements.txt
amqp==5.3.1
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
distlib==0.3.9
Django==5.1.3
django-celery-beat==2.7.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
filelock==3.16.1
flake8==7.1.1
identify==2.6.2
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
Markdown==3.7
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
prompt_toolkit==3.0.48
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.3
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.27.1
watchfiles==0.24.0
wcwidth==0.2.13
